requirements.txt
Django==4.2.7
Pillow==10.1.0
django-admin-sortable2==2.1.10
django-tinymce==3.6.1
environs==9.5.0
requests==2.31.0
